LPS-69139 Extract base class
